+++ src/Expression.h
@@ -5,13 +5,13 @@
 
 struct Expression {
   private:
-    int (&m_regular)[4];
-    int (&m_blinking)[4];
+    byte (&m_regular)[4];
+    byte (&m_blinking)[4];
   public:
-    Expression(int (&regular)[4], int (&blinking)[4]);
-    Expression(int (*regular)[4], int (*blinking)[4]);
-    int* getRegular();
-    int* getBlinking();
+    Expression(byte (&regular)[4], byte (&blinking)[4]);
+    Expression(byte (*regular)[4], byte (*blinking)[4]);
+    byte* getRegular();
+    byte* getBlinking();
 };
 
 #endif
